Splitting out the peft parameter extraction from the standard extraction to avoid deepspeed errors on GPU
